Redoing body parts
Scrapped the old way to make body parts, it's a bit restrictive and not
very nicely implemented. With the new one, it can better express how
body parts are related and how they work. Has better functionality too
for less confusion.

So, just rewriting current body parts, and going to add the rest of the
missing body parts after.

I've also moved around some of the enumeration classes from the engine
to the game folder under "Enumerations". It was just more logical that
way since the enumerations are a product of the game, not the framework.